cyslicer/requirements.txt
numpy
trimesh
numpy-stl
pymeshfix
pyvista
pyvistaqt
PyQt5
plotly
